pyproject.toml
[tool.poetry]
name = "stocks-api"
version = "0.1.0"
description = ""
authors = ["Morgan Nicholson <55922364+nichmorgan@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
inflection = "^0.5.1"
kink = "^0.8.0"
pydantic-settings = "^2.3.4"
pyjwt = "^2.8.0"
asyncpg = "^0.29.0"
sqlmodel = {version = "^0.0.19", extras = ["asyncio"]}
aiohttp = "^3.9.5"
python-dateutil = "^2.9.0.post0"
sqlalchemy-utils = "^0.41.2"
fastapi-jwt = "^0.3.0"
retry = "^0.9.2"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.1"
pytest = "^8.2.2"
types-requests = "^2.32.0.20240622"
pytest-asyncio = "^0.23.7"
types-python-dateutil = "^2.9.0.20240316"
polyfactory = "^2.16.0"
types-retry = "^0.9.9.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
